+++ src/lib/output/themes/lib.tsx
@@ -117,7 +117,10 @@ export function camelToTitleCase(text: string) {
     return text.substring(0, 1).toUpperCase() + text.substring(1).replace(/[a-z][A-Z]/g, (x) => `${x[0]} ${x[1]}`);
 }
 
-export function displayPartsToMarkdown(parts: CommentDisplayPart[], urlTo: DefaultThemeRenderContext["urlTo"]) {
+export function displayPartsToMarkdown(
+    parts: readonly CommentDisplayPart[],
+    urlTo: DefaultThemeRenderContext["urlTo"]
+) {
     const result: string[] = [];
 
     for (const part of parts) {