+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 -   Added support for using the comment directly before a constructor parameter that declares a property as the property comment, #2019.
 -   Improved schema generation to give better autocomplete for the `sort` option.
 -   Optional properties are now visually distinguished in the index/sidebar by rendering `prop` as `prop?`, #2023.
+-   `DefaultThemeRenderContext.markdown` now also accepts a `CommentDisplayPart[]` for rendering, #2004.
 
 ### Bug Fixes